+++ sdk/communication/azure-communication-administration/CHANGELOG.md
@@ -5,7 +5,7 @@
 - Added support for Azure Active Directory Authentication.
 
 ### Breaking Changes
-- CommunicationIdentityClient is moved to a new package, `Azure.Communication.Identity`.
+- CommunicationIdentityClient and CommunicationIdentityAsyncClient is moved to a new package, `Azure.Communication.Identity`.
 
 ## 1.0.0-beta.3 (2020-11-16)
 ### Added